Resolve print binary one line bug

+++ src/my-types.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-using BitType = uint32_t;	                        // Type alias, can change Bit size later
+using BitType = uint8_t;	                        // Type alias, can change Bit size later
 using MaxBitType = uint64_t;						// Max Bit Type for calculator
 constexpr int BitTypeBits = sizeof(BitType) * 8;    // Automatically calculates bit width
